refactor styles for edit modal
