+++ src/client/components/Admin/RefactorCities/AddCities.jsx
@@ -9,8 +9,7 @@ import LinkButton from '../../LinkButton/LinkButton.jsx'
 import { addCityToDB } from '../../../actions'
 import { required } from '../../../validation'
 
-import '../../../style/global-style/refactor-modal.less'
-import './RefactorCities.less'
+import '../../../style/refactor-modal.less'
 
 class AddCities extends React.Component {
   render () {
@@ -23,14 +22,14 @@ class AddCities extends React.Component {
     return (
 
       ReactDOM.createPortal(
-        <div className='modal-window modal-window-for-refactor'>
-          <div className='refactor-city basic-style-modal-refactor add-city'>
-            <div className="refactor-city__header basic-style-header">
+        <div className='modal-window-for-refactor'>
+          <div className='refactor-model'>
+            <div className='refactor-model__header'>
               Add city
-              <LinkButton to='/admin/cities' name='&times;' className='refactor-city__header__right-button-close'/>
+              <LinkButton to='/admin/cities' name='&times;' className='refactor-model__header__right-button-close'/>
             </div>
             <form
-              className='form-for-refactor'
+              className='refactor-model__form'
               onSubmit={handleSubmit(addCity)}>
               <Field
                 label='Enter city'
@@ -41,10 +40,9 @@ class AddCities extends React.Component {
                 placeholder='Enter new city'
               />
               <button
-                className='basic-style-button'
                 type='submit'
                 label='submit'
-                className='refactor-city__submit'>Submit</button>
+                className='refactor-model__form__button-submit'>Submit</button>
             </form>
           </div>
         </div>
