+++ src/client/components/Admin/RefactorCities/EditCities.jsx
@@ -11,8 +11,7 @@ import Preloader from '../../App/Preloader/Preloader.jsx'
 import { editCityIntoDB } from '../../../actions'
 import { required } from '../../../validation'
 
-import '../../../style/global-style/refactor-modal.less'
-import './RefactorCities.less'
+import '../../../style/refactor-modal.less'
 
 class EditCities extends React.Component {
   state = {
@@ -41,14 +40,14 @@ class EditCities extends React.Component {
 
     return (
       ReactDOM.createPortal(
-        <div className='modal-window modal-window-for-refactor'>
-          <div className='refactor-city basic-style-modal-refactor'>
-            <div className="refactor-city__header basic-style-header">
+        <div className='modal-window-for-refactor'>
+          <div className='refactor-model'>
+            <div className='refactor-model__header'>
               Edit city
-              <LinkButton to='/admin/cities' name='&times;' className='refactor-city__header__right-button-close'/>
+              <LinkButton to='/admin/cities' name='&times;' className='refactor-model__header__right-button-close'/>
             </div>
             <form
-              className='form-for-refactor'
+              className='refactor-model__form'
               onSubmit={handleSubmit(editCity)}>
               <Field
                 label={`ID: ${arr[arr.length - 1]}`}
@@ -65,7 +64,7 @@ class EditCities extends React.Component {
                 type='text'
               />
               <button
-                className='basic-style-button'
+                className='refactor-model__form__button-submit'
                 type='submit'
                 label='submit'>Submit</button>
             </form>
