+++ src/client/components/Admin/RefactorOrders/AddOrder.jsx
@@ -9,11 +9,10 @@ import SelectField from '../../ComponentMaterial/SelectField/'
 import DateField from '../../ComponentMaterial/DateField/'
 import LinkButton from '../../LinkButton/LinkButton.jsx'
 import { addOrdersToDB } from '../../../actions'
+import { getDate } from '../../App/OrderForm/logic'
 import { required } from '../../../validation'
 
-import '../../../style/global-style/refactor-modal.less'
-import './RefactorOrders.less'
-import { getDate } from '../../App/OrderForm/logic'
+import '../../../style/refactor-modal.less'
 
 class AddOrder extends React.Component {
   state = {
@@ -57,14 +56,14 @@ class AddOrder extends React.Component {
     return (
 
       ReactDOM.createPortal(
-        <div className='modal-window modal-window-for-refactor'>
-          <div className='refactor-orders basic-style-modal-refactor'>
-            <div className='refactor-orders__header basic-style-header'>
+        <div className='modal-window-for-refactor'>
+          <div className='refactor-model'>
+            <div className='refactor-model__header'>
               Add Order
-              <LinkButton to='/admin/orders' name='&times;' className='refactor-orders__header__right-button-close'/>
+              <LinkButton to='/admin/orders' name='&times;' className='refactor-model__header__right-button-close'/>
             </div>
             <form
-              className='form-for-refactor'
+              className='refactor-model__form'
               onSubmit={handleSubmit(addOrder)}>
               <Field
                 name='customerId'
@@ -151,7 +150,7 @@ class AddOrder extends React.Component {
                 }
               </Field>
               <button
-                className='basic-style-button'
+                className='refactor-model__form__button-submit'
                 type='submit'
                 label='submit'>Submit
               </button>
