+++ src/client/components/Admin/RefactorOrders/EditOrder.jsx
@@ -5,6 +5,7 @@ import { Field, initialize, reduxForm } from 'redux-form'
 import { Redirect } from 'react-router-dom'
 import { getDate } from '../../App/OrderForm/logic'
 import MenuItem from '@material-ui/core/MenuItem'
+import axios from 'axios'
 
 import TextField from '../../ComponentMaterial/TextField/'
 import SelectField from '../../ComponentMaterial/SelectField/'
@@ -14,9 +15,7 @@ import Preloader from '../../App/Preloader/Preloader.jsx'
 import { editOrdersIntoDB } from '../../../actions'
 import { required } from '../../../validation'
 
-import '../../../style/global-style/refactor-modal.less'
-import './RefactorOrders.less'
-import axios from 'axios'
+import '../../../style/refactor-modal.less'
 
 class EditOrder extends React.Component {
   state = {
@@ -69,14 +68,14 @@ class EditOrder extends React.Component {
     return (
 
       ReactDOM.createPortal(
-        <div className='modal-window modal-window-for-refactor'>
-          <div className='refactor-orders basic-style-modal-refactor edit-order'>
-            <div className="refactor-orders__header basic-style-header">
+        <div className='modal-window-for-refactor'>
+          <div className='refactor-model'>
+            <div className='refactor-model__header'>
               Edit Order
-              <LinkButton to='/admin/orders' name='&times;' className='refactor-orders__header__right-button-close'/>
+              <LinkButton to='/admin/orders' name='&times;' className='refactor-model__header__right-button-close'/>
             </div>
             <form
-              className='form-for-refactor'
+              className='refactor-model__form'
               onSubmit={handleSubmit(editOrder)}>
               <Field
                 label={`ID: ${arr[arr.length - 1]}`}
@@ -166,7 +165,7 @@ class EditOrder extends React.Component {
                 }
               </Field>
               <button
-                className='basic-style-button'
+                className='refactor-model__form__button-submit'
                 type='submit'
                 label='submit'>Submit
               </button>
