+++ src/client/components/Admin/RefactorWorkers/EditWorkers.jsx
@@ -4,6 +4,7 @@ import { connect } from 'react-redux'
 import { Field, initialize, reduxForm } from 'redux-form'
 import { Redirect } from 'react-router-dom'
 import MenuItem from '@material-ui/core/MenuItem'
+import axios from 'axios'
 
 import TextField from '../../ComponentMaterial/TextField/'
 import SelectField from '../../ComponentMaterial/SelectField/'
@@ -12,9 +13,7 @@ import Preloader from '../../App/Preloader/Preloader.jsx'
 import { editMastersIntoDB } from '../../../actions'
 import { required } from '../../../validation'
 
-import '../../../style/global-style/refactor-modal.less'
-import './RefactorWorkers.less'
-import axios from 'axios'
+import '../../../style/refactor-modal.less'
 
 class EditWorkers extends React.Component {
   state = {
@@ -47,14 +46,14 @@ class EditWorkers extends React.Component {
     return (
 
       ReactDOM.createPortal(
-        <div className='modal-window modal-window-for-refactor'>
-          <div className='refactor-workers basic-style-modal-refactor'>
-            <div className="refactor-workers__header basic-style-header">
+        <div className='modal-window-for-refactor'>
+          <div className='refactor-model'>
+            <div className='refactor-model__header'>
               Edit Worker
-              <LinkButton to='/admin/workers' name='&times;' className='refactor-workers__header__right-button-close'/>
+              <LinkButton to='/admin/workers' name='&times;' className='refactor-model__header__right-button-close'/>
             </div>
             <form
-              className='form-for-refactor'
+              className='refactor-model__form'
               onSubmit={handleSubmit(editWorker)}>
               <Field
                 label={`ID: ${arr[arr.length - 1]}`}
@@ -100,7 +99,7 @@ class EditWorkers extends React.Component {
                 <MenuItem key={5} value={5}>5</MenuItem>
               </Field>
               <button
-                className='basic-style-button'
+                className='refactor-model__form__button-submit'
                 type='submit'
                 label='submit'>Submit
               </button>
