--- src/client/components/Admin/RefactorWorkers/RefactorWorkers.less
@@ -1,54 +1,0 @@
-@import '../../../style/global-variable.less';
-
-.refactor-workers {
-  top: 13vh;
-  width: 340px;
-  height: 430px;
- 
-
-  &__header {
-
-    a > &__right-button-close {
-      padding: 0;
-      border: none;
-      color: inherit;
-      cursor: pointer;
-      background-color: inherit;
-      outline: none;
-      position: absolute;
-      height: 30px;
-      width: 30px;
-      top: 15px;
-      right: 15px;
-      font-size: 30px;
-      color: black;
-    }
-
-    a > &__right-button-close:active {
-      transform: translateY(2px);
-      filter: saturate(150%);
-      color: #5d0000;
-    }
-  }
-
-  form {
-
-    & > button {
-      height: 35px;
-      margin-top: 25px;
-      width: 110px;
-      font-size: 15px;
-      border: 1px solid #b9b9b5;
-      border-radius: 10px;
-      background-color: rgba(236, 236, 232, 0.8);
-    }
-
-    & > button:hover {
-      background-color: rgba(226, 226, 222, 0.8);
-    }
-  }
-}
-
-.add-worker {
-  height: 370px;
-}
