--- src/client/style/global-style/refactor-modal.less
@@ -1,49 +1,0 @@
-@import '../global-variable.less';
-
-.basic-style-button {
-    padding: 0;
-    border: none;
-    color: inherit;
-    cursor: pointer;
-    background-color: inherit;
-    outline: none;
-}
-
-.basic-style-modal-refactor {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    position: relative;
-    background-color: @modal-background-color;
-    border: 2px solid #b5b5b5;
-    border-radius: 20px;
-    box-shadow: @app-box-shadows;
-    margin: auto;
-    z-index: 2;
-}
-
-.modal-window-for-refactor {
-    position: absolute;
-    top: 0;
-    height: 100vh;
-    width: 100vw;
-    background-color: @modal-background;
-    z-index: 2;
-}
-
-.basic-style-header {
-    text-align: center;
-    width: 100%;
-    padding-top: 40px;
-    margin-bottom: 15px;
-    font-size: 20px;
-    position: relative;
-}
-
-.form-for-refactor {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
