+++ src/client/style/model-tables.less
@@ -1,4 +1,4 @@
-@import '../global-variable.less';
+@import './global-variable.less';
 
 .root-table {
     display: flex;
