+++ src/client/style/refactor-modal.less
@@ -1,0 +1,148 @@
+@import './global-variable.less';
+
+.modal-window-for-refactor {
+    position: absolute;
+    top: 0;
+    height: 100vh;
+    width: 100vw;
+    display: flex;
+    background-color: @modal-background;
+    z-index: 2;
+}
+
+.refactor-model {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: @modal-background-color;
+    border: 2px solid #b5b5b5;
+    border-radius: 20px;
+    box-shadow: @app-box-shadows;
+    width: 340px;
+    z-index: 2;
+    margin: auto;
+
+    &__header {
+        text-align: center;
+        width: 100%;
+        padding-top: 40px;
+        margin-bottom: 15px;
+        font-size: 20px;
+        position: relative;
+
+        &__right-button-close {
+            padding: 0;
+            border: none;
+            color: inherit;
+            cursor: pointer;
+            background-color: inherit;
+            outline: none;
+            position: absolute;
+            height: 30px;
+            width: 30px;
+            top: 15px;
+            right: 15px;
+            font-size: 30px;
+            color: black;
+
+            &:active {
+                transform: translateY(2px);
+                filter: saturate(150%);
+                color: #5d0000;
+            }
+        }
+    }
+
+    &__form {
+        width: 100%;
+        height: 100%;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        padding-bottom: 40px;
+
+        &__button-submit {
+            padding: 0;
+            border: none;
+            color: inherit;
+            cursor: pointer;
+            background-color: inherit;
+            outline: none;
+            height: 35px;
+            margin-top: 25px;
+            width: 110px;
+            font-size: 15px;
+            border: 1px solid #b9b9b5;
+            border-radius: 10px;
+            background-color: rgba(236, 236, 232, 0.8);
+
+            &:hover {
+                background-color: rgba(226, 226, 222, 0.8);
+            }
+        }
+    }
+}
+
+.basic-style-button {
+    padding: 0;
+    border: none;
+    color: inherit;
+    cursor: pointer;
+    background-color: inherit;
+    outline: none;
+}
+
+.basic-style-modal-refactor {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    position: relative;
+    background-color: @modal-background-color;
+    border: 2px solid #b5b5b5;
+    border-radius: 20px;
+    box-shadow: @app-box-shadows;
+    margin: auto;
+    z-index: 2;
+}
+
+
+
+.basic-style-header {
+    text-align: center;
+    width: 100%;
+    padding-top: 40px;
+    margin-bottom: 15px;
+    font-size: 20px;
+    position: relative;
+
+    .refactor-city__header__right-button-close {
+        padding: 0;
+        border: none;
+        color: inherit;
+        cursor: pointer;
+        background-color: inherit;
+        outline: none;
+        position: absolute;
+        height: 30px;
+        width: 30px;
+        top: 15px;
+        right: 15px;
+        font-size: 30px;
+        color: black;
+      }
+  
+      .refactor-city__header__right-button-close:active {
+        transform: translateY(2px);
+        filter: saturate(150%);
+        color: #5d0000;
+      }
+}
+
+.form-for-refactor {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
